authorization task

+++ nestjs-task-management/src/tasks/tasks.service.ts
@@ -5,6 +5,8 @@ import { Task } from './task.entity';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { TaskStatus } from './task-status.enum';
 import { GetTaskFilterDto } from './dto/get-tasks-filter.dto';
+import { User } from 'src/auth/user.entity';
+import { userInfo } from 'os';
 
 @Injectable()
 export class TasksService {
@@ -16,8 +18,10 @@ export class TasksService {
         private taskRepository: TaskRepository,
     ) {}
 //     private tasks:Task[] = [];
-async getTasks(filterDto:GetTaskFilterDto):Promise<Task[]>{
-return await this.taskRepository.getTasks(filterDto);
+async getTasks(filterDto:GetTaskFilterDto,
+    user:User
+    ):Promise<Task[]>{
+return await this.taskRepository.getTasks(filterDto,user);
 }
 
 //     getAllTasks():Task[]{
@@ -41,8 +45,8 @@ return await this.taskRepository.getTasks(filterDto);
 //         return tasks;
 //     }
 
-async getTaskById(id:number) : Promise<Task>{
-    const found = await this.taskRepository.findOne(id);
+async getTaskById(id:number,user:User) : Promise<Task>{
+    const found = await this.taskRepository.findOne({where:{id, userId: user.id}});
     if(!found){
         throw new NotFoundException(`Task with ID "${id}" not found`);
             }
@@ -56,19 +60,26 @@ async getTaskById(id:number) : Promise<Task>{
 //     }
 //     return found;
 //     }
-async createTask(createTaskDto:CreateTaskDto) :Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto);
+async createTask(createTaskDto:CreateTaskDto,user:User) :Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto,user);
 }
 
-        async deleteTask(id:number):Promise<void>{
-            const result = await this.taskRepository.delete(id);
+        async deleteTask(
+            id:number,
+            user:User
+            ):Promise<void>{
+            const result = await this.taskRepository.delete({id, userId: user.id});
             if(!result.affected){
                 throw new NotFoundException(`Task with ID "${id}" not found`);
             }
         }
 
-        async updateTaskStatus(id:number,status:TaskStatus):Promise<Task>{
-            const task = await this.getTaskById(id);
+        async updateTaskStatus(
+            id:number,
+            status:TaskStatus,
+            user:User
+            ):Promise<Task>{
+            const task = await this.getTaskById(id,user);
             task.status = status;
             task.save();
             return task;
